Revert "[binder] improve binder_transact()"
This caused hard shutdowns when attemting to copy large bodies of text
in Google Chrome. I suspect this is also the cause of a lot of the
other bug reports that I have gotten involving failure to allocate
binder space. Reading the commit message more carefully, it seems
like that patch needs something in either frameworks/native, art,
or system/core to be successful.

ramoops: https://hastebin.com/jihiboqefa.go

This reverts commit 5879b553ba870394dfc0659ca435ba6eb8cb67ab.

Change-Id: I3ed989adf159bc9ad756c43e1989223dd524887c
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>